♻️refactor: Reorganise supabase interacting files to lib folder not utils
