+++ app/(tabs)/profile.tsx
@@ -2,8 +2,8 @@ import OwnProfileHeader from '@/components/OwnProfileHeader';
 import PostPreview from '@/components/PostPreview';
 import Header from '@/components/TopLevelHeader';
 import { useAuth } from '@/providers/AuthProvider';
-import SupabasePostEndpoint from '@/utils/supabase/PostEndpoint';
-import SupabaseUserEndpoint from '@/utils/supabase/UserEndpoint';
+import SupabasePostEndpoint from '@/lib/supabase/PostEndpoint';
+import SupabaseUserEndpoint from '@/lib/supabase/UserEndpoint';
 import { TPost, TProfile } from '@/utils/types';
 import { useFocusEffect } from 'expo-router';
 import { useCallback, useState } from 'react';
