+++ app/post.tsx
@@ -4,9 +4,9 @@ import Tag from '@/components/Tag';
 import Header from '@/components/TopLevelHeader';
 import { useAuth } from '@/providers/AuthProvider';
 import { showToast } from '@/utils/helpers';
-import { supabase } from '@/utils/supabase/client';
-import SupabasePostEndpoint from '@/utils/supabase/PostEndpoint';
-import SupabaseUserPostInteractionEndpoint from '@/utils/supabase/UserPostInteractionEndpoint';
+import { supabase } from '@/lib/supabase/client';
+import SupabasePostEndpoint from '@/lib/supabase/PostEndpoint';
+import SupabaseUserPostInteractionEndpoint from '@/lib/supabase/UserPostInteractionEndpoint';
 import { TCommentsAndAuthors, TGetHomePagePost, TProfile } from '@/utils/types';
 import Ionicons from '@expo/vector-icons/Ionicons';
 import { useLocalSearchParams, useRouter } from 'expo-router';
